requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
et-xmlfile==1.0.1
importlib-metadata==0.19
jdcal==1.4.1
more-itertools==4.2.0
openpyxl==2.6.2
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-rerunfailures==7.0
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
yagmail==0.11.220
zipp==0.5.2
allure-pytest ==2.8.6
allure-python-commons ==2.8.6 
